# src/AutoAdaptativeSystem/AdaptionLoop/Knowledge/Knowledge.Service-OpenAPISpec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Knowledge Service",
    "description": "Demonstrates all the existing operations to access and manage Knowledge properties.",
    "version": "v1"
  },
  "paths": {
    "/Property/{propertyName}": {
      "get": {
        "tags": [
          "Property"
        ],
        "summary": "Gets a property given its name.",
        "parameters": [
          {
            "name": "propertyName",
            "in": "path",
            "description": "The name of the property to find.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The property was found. Returns the value of the property.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PropertyDTO"
                }
              }
            }
          },
          "404": {
            "description": "The property was not found.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "There was an error with the provided arguments.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Property"
        ],
        "summary": "Sets value of a given property. If the property does not exist, it will be created.",
        "parameters": [
          {
            "name": "propertyName",
            "in": "path",
            "description": "The name of the property to set.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "The DTO containing the value to set.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPropertyDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPropertyDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SetPropertyDTO"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "The property was updated or created successfully."
          },
          "400": {
            "description": "There was an error with the provided arguments.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Property"
        ],
        "summary": "Deletes the value of a given property.",
        "parameters": [
          {
            "name": "propertyName",
            "in": "path",
            "description": "The name of the property to delete.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The property was deleted successfully."
          },
          "400": {
            "description": "There was an error with the provided arguments.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "The property was not found.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/Service/{serviceName}/configuration/{configurationName}": {
      "get": {
        "tags": [
          "Service"
        ],
        "summary": "Gets a configuration property given its name.",
        "parameters": [
          {
            "name": "serviceName",
            "in": "path",
            "description": "The name of the service whose configuration we are looking for.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "configurationName",
            "in": "path",
            "description": "The name of the configuration property to find.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The configuration property was found. Returns the value of the property.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigurationDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigurationDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigurationDTO"
                }
              }
            }
          },
          "400": {
            "description": "There was an error with the provided arguments.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "The configuration property was not found.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ConfigurationDTO": {
        "type": "object",
        "properties": {
          "serviceName": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "string",
            "nullable": true
          },
          "lastModification": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "PropertyDTO": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "nullable": true
          },
          "lastModification": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "SetPropertyDTO": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}